add ReservaRestController

+++ hootel/src/main/java/com/grupo11/hootel/dao/ReservaDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ReservaDAO {
 
-    Reserva lerReservaPin(Integer pin);
+    Reserva lerReservaPin(long pin);
 
     List<Reserva> lerTodasReservas();
